refactor(tui): extract cwd-compute/session mixins from tui_gateway/server.py (shard s1)
